+++ convalida-compiler/src/main/java/convalida/compiler/model/ValidationClass.java
@@ -1,4 +1,4 @@
-package convalida.compiler.internal;
+package convalida.compiler.model;
 
 import com.squareup.javapoet.TypeName;
