+++ convalida-compiler/src/main/java/convalida/compiler/util/Messager.java
@@ -1,4 +1,4 @@
-package convalida.compiler;
+package convalida.compiler.util;
 
 import java.io.PrintWriter;
 import java.io.StringWriter;
@@ -10,24 +10,24 @@ import javax.tools.Diagnostic;
 /**
  * @author wellingtoncosta on 02/04/18
  */
-class Messager {
+public class Messager {
 
     // Can not be instantiated
     private Messager() { }
 
     private static javax.annotation.processing.Messager messager;
 
-    static void init(javax.annotation.processing.Messager messager) {
+    public static void init(javax.annotation.processing.Messager messager) {
         Messager.messager = messager;
     }
 
-    static void logParsingError(Element element, Class<? extends Annotation> annotation, Exception e) {
+    public static void logParsingError(Element element, Class<? extends Annotation> annotation, Exception e) {
         StringWriter stackTrace = new StringWriter();
         e.printStackTrace(new PrintWriter(stackTrace));
         error(element, "Unable to parse @%s validation.\n\n%s", annotation.getSimpleName(), stackTrace);
     }
 
-    static void error(Element element, String message, Object... args) {
+    public static void error(Element element, String message, Object... args) {
         if (args.length > 0) {
             message = String.format(message, args);
         }
